refactor: 진료 부분 리팩토링 시작 - 컴포넌트 위치 변경
